refactor: move entity tag DTOs to separate types module and add ghost tag deletion
- Moved DTO classes related to entity tags from `model.py` to a new `types.py` module for better organization.
- Added functionality to delete tags with no associated entities in the `TagRepository` and exposed it via a new API endpoint.
- Updated imports and dependencies across modules to reflect these changes.
