+++ app/api/v1/entity_tags/service.py
@@ -4,7 +4,7 @@ It uses the repository layer to interact with the database.
 """
 
 from typing import List
-from app.api.v1.entity_tags.model import (
+from app.api.v1.entity_tags.types import (
     EntityTag,
     EntityTagCreateRequest,
     EntityTagDeleteRequest,
@@ -63,9 +63,7 @@ class EntityTagService:
         logger.debug(f"Resetting tags for entity {request.entity_id}")
 
         logger.debug(f"Deleting existing tags for entity: {request.entity_id}")
-        deleted = self.delete(
-            EntityTagDeleteRequest(entity_id=request.entity_id)
-        )
+        deleted = self.delete(EntityTagDeleteRequest(entity_id=request.entity_id))
         logger.debug(f"Deleted {deleted} tags for entity: {request.entity_id}")
 
         for tag_group_req in request.tag_groups:
